moving color map to colorPalette file
