+++ src/style/colorsPalette.js
@@ -85,4 +85,13 @@ const colorsPalette = {
   black: '#000000',
 };
 
+// For Eslint --fix
+const extraFixColorsMap = {
+  black: 'black',
+  white: 'white',
+  '#000': 'black',
+  '#fff': 'white',
+};
+
 module.exports = colorsPalette;
+module.exports.extraFixColorsMap = extraFixColorsMap;
